refactor(layering): remove tracked_state -> live_state back-edges
Move live_state/entity_columnar.rs to a top-level entity_columnar module
(it imports nothing from live_state) and relocate
load_certified_rows_at_commit from tracked_state to live_state.
