+++ packages/lix/src/entity_columnar.rs
