+++ packages/lix/src/live_state/tracked_head.rs
@@ -21,8 +21,9 @@ pub(crate) use hot::{
     HOT_COLLECTION_CONTROL_SPACE, HOT_DIFF_SPACE, HOT_FILE_SPACE, HOT_ROW_SPACE,
     HotStateTransactionCache, HotTrackedSnapshot, PACKED_CURRENT_BASE_CONTROL_SPACE,
     PACKED_CURRENT_BASE_SPACE, PACKED_CURRENT_EXCLUSIVE_SCHEMA_BASE_SPACE,
-    PackedIdentityMembership, ROOT_CURRENT_BASE_SPACE, materialize_certified_root_rows,
-    scan_certified_history_rows, stage_certified_entity_batches, stage_retire_hot_generation,
+    PackedIdentityMembership, ROOT_CURRENT_BASE_SPACE, load_certified_rows_at_commit,
+    materialize_certified_root_rows, scan_certified_history_rows, stage_certified_entity_batches,
+    stage_retire_hot_generation,
 };
 
 /// Stable physical address of a row in an immutable columnar base.
