+++ packages/lix/src/tracked_state/storage.rs
@@ -1007,7 +1007,7 @@ async fn load_current_state_values_from_descriptors(
         }
     }
     for (&source_id, manifest) in &columnar_manifests {
-        let identity_column_index = crate::live_state::entity_identity_column_index(manifest)
+        let identity_column_index = crate::entity_columnar::entity_identity_column_index(manifest)
             .ok_or_else(|| {
                 replacement_payload_error("current-state columnar identity contract drifted")
             })?;
@@ -1062,7 +1062,7 @@ fn apply_columnar_identity_page(
     let first_key = decode_key(&descriptor.first_key)?;
     if manifest.content_digest()? != descriptor.content_digest
         || manifest.namespace != first_key.schema_key
-        || crate::live_state::entity_row_group_set_id(
+        || crate::entity_columnar::entity_row_group_set_id(
             CommitId::new(uuid::Uuid::from_bytes(descriptor.owner_commit_id)),
             &manifest.namespace,
         )
@@ -3259,7 +3259,7 @@ async fn load_scoped_current_state_descriptor_rows(
             let schema_key = decode_key(&descriptor.first_key)?.schema_key;
             if manifest.content_digest()? != descriptor.content_digest
                 || manifest.namespace != schema_key
-                || crate::live_state::entity_row_group_set_id(
+                || crate::entity_columnar::entity_row_group_set_id(
                     CommitId::new(uuid::Uuid::from_bytes(descriptor.owner_commit_id)),
                     &manifest.namespace,
                 )
@@ -6845,11 +6845,11 @@ pub(crate) fn validate_columnar_mutation_manifest(
     parts: &crate::tracked_state::types::ColumnarMutationPartSet,
 ) -> Result<(), LixError> {
     let owner = CommitId::new(uuid::Uuid::from_bytes(parts.owner_commit_id));
-    if crate::live_state::entity_row_group_set_id(owner, &parts.schema_key).as_bytes()
+    if crate::entity_columnar::entity_row_group_set_id(owner, &parts.schema_key).as_bytes()
         != parts.row_group_set_id
         || manifest.content_digest()? != parts.manifest_digest
         || manifest.namespace != parts.schema_key
-        || crate::live_state::entity_identity_column_index(manifest).is_none()
+        || crate::entity_columnar::entity_identity_column_index(manifest).is_none()
         || manifest
             .groups
             .iter()
@@ -10479,7 +10479,7 @@ pub(crate) async fn stage_delete_commit_state_manifest_for_gc(
                 format!("retired columnar mutation authority '{commit_id}' names owner '{owner}'"),
             ));
         }
-        let row_group_id = crate::live_state::entity_row_group_set_id(commit_id, &parts.schema_key);
+        let row_group_id = crate::entity_columnar::entity_row_group_set_id(commit_id, &parts.schema_key);
         if row_group_id.as_bytes() != parts.row_group_set_id {
             return Err(LixError::new(
                 LixError::CODE_INTERNAL_ERROR,
@@ -13223,7 +13223,7 @@ fn validate_commit_state_mutation_inventory(
             })
             || columnar.owner_commit_id != *commit_id.as_uuid().as_bytes()
             || columnar.row_group_set_id
-                != crate::live_state::entity_row_group_set_id(commit_id, &columnar.schema_key)
+                != crate::entity_columnar::entity_row_group_set_id(commit_id, &columnar.schema_key)
                     .as_bytes()
             || columnar.manifest_digest == [0; 32]
             || columnar.schema_key.is_empty()
